requirements.txt
# VIGIL — Trust Intelligence for the Agent Economy
# Core dependencies

# Async HTTP client (API calls to Basescan, DexScreener, Twitter, Telegram)
httpx>=0.25.0

# Web3 interaction (Base chain RPC, contract calls)
web3>=6.0.0

# Twitter API integration
tweepy>=4.14.0

# Telegram bot
python-telegram-bot>=20.0

# Environment variable management
python-dotenv>=1.0.0

# Data handling
pydantic>=2.0.0

# Optional: GAME SDK integration (install separately)
# pip install game-python
